fix: get loading state everywhere
